Test of new commit emails
Hi Joel, this is a test of the new svn commit emails. The first line you 
write in the message is the subject of the email.

Sorry about the deluge of commit emails previously. There should not be 
any more. (Except, of course, for real commits.)

You won't see it in this email, but the diffs now have strike-throughs 
and underlines to indicate deletions/additions repsectively.