Proposed timetable
I added some deadlines to the TODO file. Feel free to modify as you deem 
fit. Note that the file doesn't reflect when we will be doing stuff, but 
rather when stuff will be done by. For example, I have put 
proof of termination in week 7, but obviously it will take longer than 
one week to do that.