Can declare and initialize vars on same line.
Doesn't yet work for for loops, and I'm still putting in the locations. I'm committing it because the yices stuff 
doesn't seem to be working anymore, and I'm not sure why. (Even when I comment out all of my changes, it still doesn't 
work.) Joel, do you know what the problem is?

- Jason