TODOs
Moved the client generate xml TODO up to week 3, because we need this in 
order for the client to communicate whether generate runtime 
assertions should be generated.

Also added some TODOs.

Joel, it's probably easiest if you generate the xml on the client side 
first,  and then when you're done with that I immediately add code to 
parse it on the server side. Does that sound ok to you?