Arrays DP mostly working
The arrays DP code appears to be working. The programs Max, LinearSearch, and BubbleSort all verify correctly.
(BubbleSort is a little annoying, because I had to write out the full versions of the sorted and partitioned predicates.)

Unfortunately, the VCs that are fed to Yices on BubbleSort are very, very long and sometimes it takes yices >10 seconds to decide sat/unsat.
I know why this is happening - it's to do with the way I'm applying the arrays DP. I've added a TODO to return to my code and
make it generate smaller VCs.

Anyway, provided you stay within the arrays DP fragment (basically, never have an index of an array read/write involve an element of an array e.g. arr[arr2[j]]
and you'll be fine), I think it should work for everything. i.e. there are no bugs I'm aware of.

A few more things:
1. There was a bug in your yices generation code to do with unary minus, but I fixed it.
2. I don't like the idea of ignoring the "inconsistant state" error message that yices sometimes returns. That error message was important (it was being given
   in response to point number 1).
   I changed it to assert(false). Perhaps you know more than I do about why you were ignoring that, in which case let me know. Otherwise, I think
   we should treat it as a serious error and assert/throw an exception.
3. Is it possible to make the calls to yices concurrently?
4. Your caching code works inconcistently. Sometimes it seems to work, but mostly not. It's possible that the huge length of the VCs generated by BubbleSort
   is screwing it up.