Predicates and other stuff
1. Predicates are now working. There is a file "included.pi" which is 
currently set to go in the same directory as the executable. (there's a 
sonstant in constants.ml) You can also declare predicates in any file 
you want.

2. I figured out how to get the directory of the executable - it's the 
0th argument passed to the program. I created a function in utils.ml to 
get an absolute path to a file, given a path relative to the executable. 
I changed the yices path to use this, so we can now run the program from 
any working directory.

3. piVC is now set up to use multiple files in a single compliation. I 
had to do this to get included.pi to work. I doubt we'll ever use this 
functionality any more than it's already being used, but we have it.