Parser/lexer support for termination arguments.
For parser simplicity, atm termination arguments need to go *before* 
the annotation(s). (This can be changed later without affecting the 
ast.)

Joel, all I did was change the productions. You'll notice that the 
productions for ForStmt, WhileStmt, an FnDecl now have an element 
OptionalTermination. This is an Expr opt. Its currently being ignore, so 
you'll need to grab it (using the dollars sign followed by its element 
number on the RHS) and stick it into the AST.

- Jason