Syntax highlighting.
You'll need to add syntax.jar to your build path.

If you're wondering why I got rid of a few of the lines in PiCode.java, 
it's because we don't actually change the document when a new file 
loads, so there's no need to re-attach the undo manager and reset the 
tab size, as was necessary before.