New config system for client
Hey Joel, this new config system is super-elegent. Every config value 
must have an entry in the DEFAULTS array of Config.java. Every time the 
program starts, it checks for the following three things:
- config files doesn't exist
- config file exists but is missing some keys
- config file exists and has extra keys which are now deprecated because 
they are no longer in the DEFAULTS array

After checking those three things, it writes a new config file if 
necessary.

Adding a new config key/value is very simple: all you do is add it to 
add it to the Config.DEFAULTS array. Every config value must be present 
in the DEFAULTS array.

Also, you've probably inferred this, but there's no default config file 
anywhere. Rather, if a config file doesn't exist, one is created from 
the DEFAULTS array. I think this is a good way to do it.

I'm a little nervous about commiting this because I'm not sure how the 
home directory stuff is going to work on Windows. Let me know how it 
goes...

Also, I added a much-needed feature: the gui remembers the last 
directory you loaded/saved a file from/to, and defaults to that in the 
future. No more needing to put .pi files in your working directory to 
save time!