The GUI sends the code in xml and the GUI receives and parses it.
Note that we ran into line ending issues again: the incoming xml had "&#13;" at the end of each line.
I had to rewrite Xml_generator.replace_bad_chars to be in the opposite direction and to remove those.
So if we eventually get rid of the conver_line_endings function, we'll want to change this too.
Also, make sure this works for you, Jason.

Lastly, I added a TODO to not allow updates to parameters (re Aaron's email about how to handle _0).