bug fix, test_parser working
1. func calls now work in if conditions. soon i'll have them working in 
all conditions (i.e. while/for), but i need to go to the gym now...

2. test_parser now works again. yay! the new makefile stuff had done 
something weird to it. Note that I had to put included.pi in the 
compilers directory, because constants.ml says to looks for that file 
in the same directory as the bin. When we get the centralized bin system
working this will no longer be necessary.