Config system on server finished.
Peruse config.mli for details. It's very simple.

I checked the actual config files into svn. That way everything works 
for you and Aaron, and new keys/values can easily be added to the config 
system.

Also, minor new feature: paths can now be either relative or absolute. 
(It is taken as absolute iff it begins with a '/' character.)

- Jason