Ensure ranking annotations are nonnegative and add termination argument support to the GUI.
Whew, this is a big commit.  I ended up having to change the main data structure a fair bit and the functions that use it a fair bit, although it's similar.  We basically pass around (fn * validity * (validity * basic paths) * termination_info).

I added lots of new nodes to the GUI tree to support this.  We'll probably want to hide or auto-expand some of them, but I'll leave this for the moment.