Jar file is fully self-contained
From now on, if you want to use a resource (i.e. image) in the jar file, use the function getURL(String location) in Utils.java. This function is jar-aware so it works both in Eclipse and 
as a jar.

All new jar-based libraries should go in the libraries directory, and all new images should go into the images directory. If you want to do anything else, you'll need to edit the Makefile.

Remember, I  want to keep the jar fully self-contained (so that the users don't need the full directory structure just to run the client). Thus, be careful never to use resources 
that aren't included in the jar file.

- Jason