Command line argument support for the servers.
Use the --help option to see a list of available options.  We can easily add more later.
Whenever you want to print something in a server only if printing is enabled, call Config.print.  If you just call print_string or print_endline, it will always print (obviously).
I changed the default truncation length to 80 so the msgs fit perfectly on a standard-sized terminal.
It recognizes nonexisting arguments and prints out a help message (although OCaml does the work for that).