JList selection bugfix/question.
I fixed the compile error assertion bug.  But do you notice the annoying behavior both the error pane and the vc pane had where if you clicked below the last element it selected the last element?  That annoyed me, so I fixed it.  Java is stupid and JList is ugly, so it's hard to do, but I did it.  If you agree with me that the default behavior is stupid, let's change the vc pane like this as well.  What do you think?