More easy GUI changes and a question.
Remember how we wanted it so that if you click on the tree outside of something it unhighlights?  Well, the only way to do it is to compare the coordinates of the click with the closest row.  But the problem is that the little show/hide children toggler isn't part of the object.  So we can't tell the difference between clicking to the left of the row with clicking on that to show the children.  But in the first case we want to unselect and the second we don't.  Right now I hard-coded the distance so it works, but that's probably platform-specific.  Can you test it, Jason?  And if it's different, what should we do?  Overestimate it, or just make it like I had before where you had to click below or to the right to unselect?