Finished termination arguments (pending no bugs found in testing) and added support for floats.
Notes:

compiler/ast_utils.ml can be used for, well, utility functions that operate on the ast somehow.

For floats, you still can't do arithmetic operations on a float and an int.  Yices supports this, I believe, but since we're doing formal methods, I feel it's best not to allow behind-the-scenes implicit conversions.

The compiler seems to be running slow on termination_example.pi for me, but then it runs fine after I restart the server, so I don't know if it's a problem or not.  Do you notice extra slowdown?  I don't otherwise.