client refactoring complete
Joel:

I think that the "names" should actually be a property of the atoms instead of the basic paths. For a correctness or decreasing atom, we can use the basic path 
name that you've generated, and for the nonnegative atoms, we can just use the tuple as the name. Does that sound good to you?

Currently, the only thing that needs to be done to get the names to show up is to add it to the constructor of VerificationAtom. Currently, I just put a dummy 
string in the constructor that says "Joel: put label here".

- Jason