Found some possible bugs.
LinearSearch-correctness.pi doesn't verify even though our solution seems to be the same as the one on the website.  Am I missing something, or is there a problem?  I have to head out right now so I can't look at it.

crash.pi, well, crashes.  The problem is with the length operator: remove that part of the annotation and it works just fine.