Some logging work.
We did assign logging to me, right?

A few questions:

Should we do any kind of automated log rotation, or should we just always append?

What do we want to log?  Obviously exceptions/errors, and I think we wanted to log all main server connections.  Is there anything else?

How should we handle the config for this?  If we do it the normal way, we'd put a log_file key in both config files.  But this would let you log to different files for the different servers.  It's probably better to always use one file and inside differentiate which is logging it (which we currently are on track to do).  So should we add a third conf file for shared things?  This could also apply to other things, such as truncate_output.