names and locations for nonnegative verification atoms
incidentally, directories need to be set to be executable in order to be accessed.
thus, the logging code wasn't working at first,  but i changed the permission
to 755 (which is the default permissions for new directories on my system)
and now it works fine. i also did something with the umask that is not
important, but was recommended on http://osdir.com/ml/lang.ocaml.beginners/2003-05/msg00044.html.
I'm just telling you this so you don't say "wtf this is weird" and remove it.