Static analysis step 1: build a cfg.
I wrote the code to build a cfg for Pi.  It probably doesn't work perfectly: I've debugged it a good bit, but not that thoroughly.  Still, it works well enough, so I wanted to commit it.  To test it out, run bin/test_cfg and give it a filename as the first and only argument (although that binary is just for debugging, so don't worry about bugs in it).

Also, as a random note, I moved the Expr_map and Expr_set to ast_utils and made them public, so they're there if you want them.