The GUI loads files passed to it on the command line.
This seems moderately useful.  It will actually launch a new window for each file passed to it, so you can pass it multiple files.  This is a bit confusing, since all are using the same java base and so if you Quit one they all quit, but they do operate independently.