Fixed two bugs.
I fixed two bugs: a problem with integer division and a compiler assert.

See the change in expr_utils.ml.  One of the matches was missing a Div match, which we didn't want.  I noticed that get_index_set in the same file is also missing it twice.  Is that correct, since real division and arrays don't mix well, or is it a mistake?