Slightly prettified yices crash output.
If you pass yices a divide by zero, it exits.  Before we asserted false.  Now we throw an exception.  It's not very nice, but I don't know how to get the error message it returns since the yices process has already exited.  We don't get the assertion line number, but we do get the error name, and it's a unique name.

I also added tests for this bug and for the div stuff from my previous commit.