Stop ci_yices from writing to .yices.log and instead capture the info ourselves.
I've been meaning to do this for a while, but I finally did it.  It has 
two advantages.  Firstly, we can use this information to get back errors 
from yices (which so far only seems to happen on a divide by zero, but 
hey).  Secondly, I think this might stop us from getting the ENFILE (too 
many open files) error.  So good times.