Timeouts.
Both the main server and the dp server time out after a specified number of seconds (given in the conf file but not the command line since there's a small trick to it).  If the dp server times out, it kills yices if needed and then throws a timeout exception back to the main server.

I've tested this on array_updates.pi and by replacing yices with a program that infinite loops.  Note that array_updates.pi times out in the main server not the dp server, so neither of these actually test a real yices timeout.  I would assume it would work fine, since the infinite loop did, but that's just a thought.  If you can figure out how to try it, go ahead.

I set the default timeouts to five and ten seconds, but once we reach production, you may want to up those a bit since load could slow things down.