support for structs
unfortunately i haven't had time to implement full support for classes. it turned out to be a bigger job than i anticipated. i'm a little nervous about bugs etc in pivc so i think it's unwise to spend this time implementing classes when there are more important things to be done. thus, i've permanently shelved support for classes. but on the bright side, structs are working nicely.