Cleanup part 1: Commiting the .depend files.
We need this since I added new files and new dependencies.  I wanted to 
commit it from Unix just to be sure.

Also, an extra note about Z3.  It gives counterexamples for arrays in 
the form a -> {0 -> 42; else -> 43}: that is, it gives a default value 
for everything else.  For simiplicity, I represented this with a a[*]=43 
counterexample node.