better naming of arrays created using new array[size] syntax
I added a new field of type "identifier option ref" to the NewArray type.
This field stores an optional preferred name for the array. If that field
is populated (which occurs during basic path construction), it uses that
identifier instead of the default _v identifier.

Note that we still need to use _v identifers if the 'new array' syntax is used
as, say, an argument to a function call.

Anyway, everything should be working as you desire now.