Possible fixed the latest error.
The cause of the bug was using the name of a function or predicate in an assertion without calling 
it (e.g. @post sorted rather than @post sorted(a, 0, 42)).  This fix catches it as a semantic 
error and all our other -soln files still work.  However, I'm a bit unsure of whether this is 
correct.  Before, we didn't throw an error when we hit these cases.  Is this correct?  At first 
thought it seems like not, but on the other hand, the name of a function or predicate is never an 
lval since you can't assign to it or have a variable holding.  You can only use them in calls, 
which don't have lvals.

I'm not going to update the server with this since I'm not quite sure if it's correct.  If you 
back me up on it, Jason, then I will.