Increased the DP server timeout to hopefully fix the bug that kills the server.
The bug that this test case causes seems to be related to timeouts.  I think that somehow the 
timeout isn't getting handled correctly and is accidentally killing the server.  This is weird, 
since it definitely worked before.  But this example seems to work with a timeout of 120s rather 
than the original 60s.  I upped it to 300s to be safe.  Hooray for workarounds.